UserService/Handlers/GetUserByIdQueryHandler.cs
using MediatR;
using UserService.Models;
using UserService.Queries;
using UserService.Data;

namespace UserService.Handlers
{
    public class GetUserByIdQueryHandler : IRequestHandler<GetUserByIdQuery, User>
    {
        private readonly IUnitOfWork _unitOfWork;

        public GetUserByIdQueryHandler(IUnitOfWork unitOfWork)
        {
            _unitOfWork = unitOfWork;
        }

        public async Task<User> Handle(GetUserByIdQuery request, CancellationToken cancellationToken)
        {
            return await _unitOfWork.Users.GetUserByIdAsync(request.Id);
        }
    }
}
